+++ src/main/java/pl/edu/pw/mini/ingreedio/api/review/repository/ReviewRepository.java
@@ -13,10 +13,10 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     Optional<Review> findById(int id);
 
     @Query("SELECT r FROM Review r WHERE r.user.id = :userId AND r.productId = :productId")
-    Optional<Review> findByUserIdAndProductId(@Param("userId") int userId,
-                                              @Param("productId") Long productId);
+    Optional<Review> findByUserIdAndProductId(@Param("userId") long userId,
+                                              @Param("productId") long productId);
 
     @Query("SELECT r FROM Review r WHERE r.productId = :productId")
-    List<Review> getProductReviews(@Param("productId") Long productId);
+    List<Review> getProductReviews(@Param("productId") long productId);
 
 }
